added special NullWrapper object to allow inserting nulls into FSink's without terminating the stream. Modified the 'Tuples' to be smart about the NullWrapper type and replace with null.
This allows the testCombineNull() test to pass, while not causing any other side effects.

The NullWrapper object scheme will work when extracting data with any of the FStream.combine() methods.  Any other extraction of data from the FStream will need to handle the NullWrapper object if nulls are expected in the stream.